+++ src/short-urls/SearchBar.js
@@ -7,23 +7,19 @@ import moment from 'moment';
 import SearchField from '../utils/SearchField';
 import Tag from '../tags/helpers/Tag';
 import DateRangeRow from '../utils/DateRangeRow';
-import { compareVersions, formatDate } from '../utils/utils';
-import { serverType } from '../servers/prop-types';
+import { formatDate } from '../utils/utils';
 import { shortUrlsListParamsType } from './reducers/shortUrlsListParams';
 import './SearchBar.scss';
 
 const propTypes = {
   listShortUrls: PropTypes.func,
   shortUrlsListParams: shortUrlsListParamsType,
-  selectedServer: serverType,
 };
 
 const dateOrUndefined = (date) => date ? moment(date) : undefined;
 
-const SearchBar = (colorGenerator) => {
-  const SearchBar = ({ listShortUrls, shortUrlsListParams, selectedServer }) => {
-    const currentServerVersion = selectedServer ? selectedServer.version : '';
-    const enableDateFiltering = !isEmpty(currentServerVersion) && compareVersions(currentServerVersion, '>=', '1.21.0');
+const SearchBar = (colorGenerator, ForServerVersion) => {
+  const SearchBar = ({ listShortUrls, shortUrlsListParams }) => {
     const selectedTags = shortUrlsListParams.tags || [];
     const setDate = (dateName) => pipe(
       formatDate(),
@@ -38,7 +34,7 @@ const SearchBar = (colorGenerator) => {
           }
         />
 
-        {enableDateFiltering && (
+        <ForServerVersion minVersion="1.21.0">
           <div className="mt-3">
             <DateRangeRow
               startDate={dateOrUndefined(shortUrlsListParams.startDate)}
@@ -47,7 +43,7 @@ const SearchBar = (colorGenerator) => {
               onEndDateChange={setDate('endDate')}
             />
           </div>
-        )}
+        </ForServerVersion>
 
         {!isEmpty(selectedTags) && (
           <h4 className="search-bar__selected-tag mt-3">